✨ Add Code Scanning Tools Integration Page
Documents how to integrate Pixeebot with Sonar. It doesn't feel right to
only have one tool on this page, so I mentioned that CodeQL is "coming
soon". Practically, we have CodeQL integration by ways of GitHub
Advanced Security right now, and documenting this is the subject of
another task we will get to next week.