🚚 Prefer the Term Preferences to Configuration (#166)
When describing the mechanism by which the user tailors Pixeebot to
their preferences, we want to use the term "preferences" instead of
"config". Preferences implies that making changes is optional: Pixeebot
still works out-of-the-box without requiring users to configure it. On
the other hand, configuration is typically associated with a required
step in the process of beginning to use a tool.

/close #work