Docs refactor and updates (#187)
This PR refactors and updates our docs to put a greater emphasis on
triage and tool remediation. It is intended as a starting point for
additional docs update. The docs are intended to answer the following
questions of a potential customer:

* What does Pixeebot do?
* What tools/languages/rules does Pixeebot support?
* How do I get started?

Probably some additional cross-links between sections would be useful.

I did not attempt to improve the styling at all; this might be worth
revisiting at a later date.

---------

Co-authored-by: Arshan Dabirsiaghi <arshan.dabirsiaghi@gmail.com>
Co-authored-by: Surag <129796490+sip49@users.noreply.github.com>